multi-node OSFT bug fixes + improved checkpointing logic
The current implementation of OSFT has calls to `distributed.get_rank()`
in places where the local rank is expected to be used. This leads to
the incorrect GPU device being used during training. This commit
fixes the problem by ensuring the correct usage of local & global rank.

Another problem is that we currently have multiple places where
checkpoints must be saved, with potentially redundant checkpoints being
saved (end of epoch, then final), but checkpointing for OSFT is very
expensive (~5 mins per checkpoint for an 8B model with solid state
storage).
This commit consolidates the logic to a single function of
`should_save_checkpoint` that makes it easy to consolidate
our complex checkpointing logic without over-engineering
the rest of the codebase.